backend/requirements.txt
fastapi
uvicorn
authlib
httpx
python-dotenv
pyjwt
itsdangerous
